requirements.txt
allure-pytest
configparser
jsonschema
PyHamcrest
pytest
pytest-rerunfailures
pytest-xdist
PyYAML
requests
selenium
webdriver-manager